Better CLI error message for build failures
