+++ dockermake/errors.py
@@ -1,3 +1,5 @@
+from __future__ import print_function
+
 # Copyright 2017 Autodesk Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,7 +13,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+from io import StringIO
+import pprint
+from termcolor import cprint
 
 class UserException(Exception):
     """
@@ -68,3 +72,20 @@ class ParsingFailure(UserException):
 
 class MultipleIgnoreError(UserException):
     CODE = 51
+
+
+class BuildError(Exception):
+    CODE = 200
+
+    def __init__(self, dockerfile, item, build_args):
+        with open('dockerfile.fail', 'w') as dff:
+            print(dockerfile, file=dff)
+        with StringIO() as stream:
+            cprint('Docker build failure', 'red', attrs=['bold'], file=stream)
+            print(u'\n   -------- Docker daemon output --------', file=stream)
+            pprint.pprint(item, stream, indent=4)
+            print(u'   -------- Arguments to client.build --------', file=stream)
+            pprint.pprint(build_args, stream, indent=4)
+            print(u'This dockerfile was written to dockerfile.fail', file=stream)
+            stream.seek(0)
+            super(BuildError, self).__init__(stream.read())
